添加购物车添加功能

+++ Milktea/src/main/java/com/milktea/common/BaseContext.java
@@ -1,13 +1,13 @@
 package com.milktea.common;
 
 public class BaseContext {
-    private static ThreadLocal<Long> threadLocal = new ThreadLocal<>();
+    private static ThreadLocal<Integer> threadLocal = new ThreadLocal<>();
 
     /**
      * 设置值
      * @param id
      */
-    public static void setCurrentId(Long id){
+    public static void setCurrentId(Integer id){
         threadLocal.set(id);
     }
 
@@ -15,7 +15,7 @@ public class BaseContext {
      * 获取值
      * @return
      */
-    public static Long getCurrentId(){
+    public static Integer getCurrentId(){
         return threadLocal.get();
     }
 }
